Added Mockito for future testing (network mocking) Restructured Project (split main file and added respective folders for future widgets) Added basic test, checking if all input fields/buttons on login are present
