+++ lib/screens/new_screen.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class NewScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('New Screen')),
+      body: const Center(
+        child: Text(
+          'This is a new screen',
+          style: TextStyle(fontSize: 24.0),
+        ),
+      ),
+    );
+  }
+}
